+++ neo/libs/libbinkdec/src/BinkDecoder.cpp
@@ -78,7 +78,8 @@ BinkHandle Bink_Open(const char* fileName)
 	classInstances.push_back(newDecoder);
 
 	// get a handle ID
-	newHandle.instanceIndex = classInstances.size() - 1;
+	// SRS - added int cast for type consistency
+	newHandle.instanceIndex = (int)classInstances.size() - 1;
 	
 	return newHandle;
 }
